Final V1.1
This. This good sir, is what happens when a print statement is what it takes to save an hour of bug fixing. I'm not saying I used it to help bug test. No. It fixed the entire program because netbeans decided it needed to wait for something to happen in a while loop.